Introduce bimodal distribution option for task wait
It is more realistic that some tasks take much more than the others.
It can happen in various reasons, and unless you have embarassingly
parallel problem, you may need to assume the time spent on each task
is unpredictable.

In order to simulate the situation, we have a flag bimordal distribution,
which makes the task wait distribution randomly multiplies by 2.